Fix all for header (#7) Move all file to it necessary place Update angular md dependance Fix for PR requests
