+++ src/app/app.component.html
@@ -1,2 +1,2 @@
-<router-outlet></router-outlet>
 <app-header></app-header>
+<router-outlet></router-outlet>
