--- src/app/components/login/login.component.spec.ts
@@ -1,3 +1,0 @@
-describe('LoginComponent', () => {
-
-});
