+++ src/app/core/app.routes.ts
@@ -1,8 +1,8 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { LoginComponent } from '../components/login/login.component';
-import { PageNotFoundComponent } from '../page-not-found/page-not-found.component';
 import { ModuleWithProviders } from '@angular/core/src/metadata/ng_module';
+import { LoginComponent } from '../components/login/login.component';
+import { PageNotFoundComponent } from '../components/page-not-found/page-not-found.component';
 import { AuthGuard } from './auth.guard';
 
 const router: Routes = [
