--- src/app/page-not-found/page-not-found.component.spec.ts
@@ -1,3 +1,0 @@
-describe('PageNotFoundComponent', () => {
-
-});
